#59 그룹 수정 이미지 표시, firebase 설정 모듈화
