+++ index.html
@@ -7,7 +7,7 @@
   <title>Sign In / Sign Up Form</title>
   <link rel="stylesheet" href="css/login.css">
   <script type="module" src="js/login.js"></script>
-
+  <script type="module" src="js/scripts.js"></script>
 </head>
 
 <body>
