+++ js/group.js
@@ -1,21 +1,5 @@
 // Firebase SDK import
-import { initializeApp } from "https://www.gstatic.com/firebasejs/9.22.0/firebase-app.js";
-import { getFirestore, doc, updateDoc, deleteDoc } from "https://www.gstatic.com/firebasejs/9.22.0/firebase-firestore.js";
-import { collection, addDoc, getDocs, query, where } from "https://www.gstatic.com/firebasejs/9.22.0/firebase-firestore.js";
-
-const firebaseConfig = {
-    apiKey: "AIzaSyA8Do_4_ZDADE64D6v1gbF36_NfaRDvh24",
-    authDomain: "ctrls-miniproject.firebaseapp.com",
-    projectId: "ctrls-miniproject",
-    storageBucket: "ctrls-miniproject.firebasestorage.app",
-    messagingSenderId: "496866464655",
-    appId: "1:496866464655:web:8d866192211cf7699d31fa",
-    measurementId: "G-SGY1JM32JF"
-};
-
-// Firebase 인스턴스 초기화
-const app = initializeApp(firebaseConfig);
-const db = getFirestore(app);
+import { collection, getDocs, doc, query, where, addDoc, updateDoc, deleteDoc, db } from './scripts.js'
 
 // Groups 불러오기
 let groupArr = [];
