+++ js/scripts.js
@@ -5,3 +5,24 @@
 */
 // This file is intentionally blank
 // Use this file to add JavaScript to your project
+
+// Firebase SDK import
+import { initializeApp } from "https://www.gstatic.com/firebasejs/9.22.0/firebase-app.js";
+import { getFirestore, collection, getDocs, doc, query, where } from "https://www.gstatic.com/firebasejs/9.22.0/firebase-firestore.js";
+import { addDoc, updateDoc, deleteDoc } from "https://www.gstatic.com/firebasejs/9.22.0/firebase-firestore.js";
+
+const firebaseConfig = {
+    apiKey: "AIzaSyA8Do_4_ZDADE64D6v1gbF36_NfaRDvh24",
+    authDomain: "ctrls-miniproject.firebaseapp.com",
+    projectId: "ctrls-miniproject",
+    storageBucket: "ctrls-miniproject.firebasestorage.app",
+    messagingSenderId: "496866464655",
+    appId: "1:496866464655:web:8d866192211cf7699d31fa",
+    measurementId: "G-SGY1JM32JF"
+};
+
+// Firebase 인스턴스 초기화
+const app = initializeApp(firebaseConfig);
+const db = getFirestore(app);
+
+export { collection, getDocs, doc, query, where, addDoc, updateDoc, deleteDoc, db };
