+++ main.html
@@ -79,7 +79,7 @@
         <!-- Bootstrap core JS-->
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"></script>
         <!-- Core theme JS-->
-        <script src="js/scripts.js"></script>
+        <script type="module" src="js/scripts.js"></script>
         <script type="module" src="js/main.js"></script>
     </body>
 </html>
